requirements.txt
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
brukva==0.0.1
certifi==2016.2.28
decorator==4.0.9
Django==1.9.5
django-appconf==1.0.2
django-compressor==2.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-libsass==0.6
django-parsley==0.6
Django-Select2==5.8.4
djangorestframework==3.3.3
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
libsass==0.11.1
Markdown==2.6.6
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
rcssmin==1.0.6
redis==2.10.5
rjsmin==1.0.12
simplegeneric==0.8.1
singledispatch==3.4.0.3
sockjs-tornado==1.0.3
tornado==4.3
tornado-redis==2.4.18
traitlets==4.2.1
